Try to skip over completely blank lines in the mapping file.

+++ meta-scripts/build_drop_import_tables_sql.py
@@ -10,7 +10,10 @@ def run(data_root, working_dir, config):
     sql_file = open("%s/drop_import_tables.sql" % (working_dir,), 'w')
 
     sqn_lookup_file = csv.DictReader(open("%s/Sequence_Number_and_Table_Number_Lookup.txt" % data_root, 'rU'))
-    for sqn, rows in groupby(sqn_lookup_file, key=lambda row: int(row[sqn_col_name])):
+    for sqn, rows in groupby(sqn_lookup_file, key=lambda row: int(row[sqn_col_name]) if row['File ID'] else None):
+        if not sqn:
+            continue
+
         sql_file.write("DROP TABLE IF EXISTS tmp_seq%04d;\n" % (sqn,))
         sql_file.write("DROP TABLE IF EXISTS tmp_seq%04d_moe;\n" % (sqn,))
